requirements.txt
openai
py-cord-dev==2.5.0rc5
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client